move common string composition functions under pkg/utils/compose
